chore: Move components to src folder
