+++ pages/_app.js
@@ -1,6 +1,6 @@
 import "@/src/styles/globals.css";
-import Layout from "@/components/layouts/Layout";
-import GoogleFontsProvider from "@/components/layouts/GoogleFontsProvider";
+import Layout from "@/src/components/layouts/Layout";
+import GoogleFontsProvider from "@/src/components/layouts/GoogleFontsProvider";
 
 
 function MyApp({ Component, pageProps }) {
